+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/Table/components/Row.tsx
@@ -160,7 +160,8 @@ const RowComponentInner = <
       className={cn(
         "group transition-colors hover:bg-f1-background-hover",
         "after:pointer-events-none after:absolute after:inset-x-0 after:bottom-0 after:h-px after:w-full after:bg-f1-border-secondary after:content-['']",
-        noBorder && "after:bg-white-100"
+        noBorder && "after:bg-white-100",
+        !!nestedRowProps?.onLoadMoreChildren && "hover:bg-transparent"
       )}
     >
       {source.selectable && (